upgrade jar version

+++ disjob-admin/ruoyi-framework/src/main/java/com/ruoyi/framework/config/ShiroConfig.java
@@ -237,7 +237,7 @@ public class ShiroConfig
         // 删除过期的session
         manager.setDeleteInvalidSessions(true);
         // 设置全局session超时时间
-        manager.setGlobalSessionTimeout(expireTime * 60 * 1000);
+        manager.setGlobalSessionTimeout((long) expireTime * 60 * 1000);
         // 去掉 JSESSIONID
         manager.setSessionIdUrlRewritingEnabled(false);
         // 定义要使用的无效的Session定时调度器
